Requirement: #700 - In calculation of the mean, the sum should be Int64, since it can easily exceed the range of an Int32.

+++ ImageViewer/Tools/Measurement/RoiStatisticsCalculator.cs
@@ -70,7 +70,7 @@ namespace ClearCanvas.ImageViewer.Tools.Measurement
 			GrayscaleImageGraphic image,
 			IsPointInRoiDelegate isPointInRoi)
 		{
-			int sum = 0;
+			Int64 sum = 0;
 			int pixelCount = 0;
 
 			image.PixelData.ForEachPixel(
